refactoring of i18n, locales, checkbox, langSwitcher, DeckPage & creating useDecks hook
