+++ src/common/locales/languages/en/translation.ts
@@ -1,4 +1,4 @@
-import { Translation } from '@/common/locales/TranslationType'
+import { Translation } from '@/common/locales/types/TranslationType'
 
 export const enTranslation: Translation = {
   cardsPage: {
