+++ src/common/locales/languages/kz/translation.json
