+++ src/common/locales/languages/ua/translation.ts
@@ -1,4 +1,4 @@
-import { Translation } from '@/common/locales/TranslationType'
+import { Translation } from '@/common/locales/types/TranslationType'
 
 export const uaTranslation: Translation = {
   cardsPage: {
