+++ src/common/locales/types/TranslationType.ts
