+++ src/common/locales/types/i18next.d.ts
@@ -1,0 +1,11 @@
+import { enTranslation } from "@/common/locales/languages/en/translation";
+import resources from "@/common/locales/types/resources";
+
+
+/*const resources2 = { translation: enTranslation } as const
+
+declare module 'i18next' {
+  interface CustomTypeOptions {
+    resources: typeof resources2
+  }
+}*/
