+++ src/common/locales/types/resources.ts
@@ -1,0 +1,7 @@
+/*
+import { enTranslation } from '@/common/locales/languages/en/translation'
+
+const resources = { translation: enTranslation } as const
+
+export default resources
+*/
