+++ src/components/TableComponent/SingleRowCard/SingleRowCard.module.scss
@@ -155,13 +155,46 @@
   }
 }
 
-.blur,
+/*.blur,
 .unBlur {
-  filter: blur(5px);
+  filter: blur(2.5px) contrast(1);
+  opacity: 0.5;
   cursor: pointer;
 }
 .unBlur {
   filter: none;
+  opacity: 1;
+}*/
+
+.blur {
+  /*filter: blur(2.5px);
+  backdrop-filter: blur(10px);
+
+
+  background: var(--color-dark-300);
+  opacity: 0.5;
+  cursor: pointer;
+  transition: opacity 0.3s ease, filter 0.3s ease;*/
+  /*content: var(--hidden);
+  filter: blur(0.5em);*/
+
+
+  text-shadow: 0 0 10px white;
+  transition: 0.1s all;
+}
+
+.blur p {
+  color: transparent;
+}
+
+.blur img {
+  filter: blur(2px);
+}
+
+.unBlur {
+  /*opacity:1;*/
+  text-shadow: none;
+  transition: 0.1s all;
 }
 
 .sell {
