+++ src/components/ui/LocaleSwitcher/LocaleMenuItem.tsx
@@ -1,21 +1,18 @@
-import { useTranslation } from 'react-i18next'
-
-import l from '@/common/locales/LangData'
-import { LangType } from '@/components/ui/LocaleSwitcher/LocaleSwitcherDrop'
+import { FullName } from '@/components/ui/LocaleSwitcher/LocaleSwitcherDrop'
 import Typography from '@/components/ui/Typography/Typography'
 import * as DropdownMenu from '@radix-ui/react-dropdown-menu'
 
 import s from './LocaleSwitcher.module.scss'
 
 type DropdownMenuItemProps = {
+  fullName: FullName
   icon: string
   isoCode: string
   onSelect: () => void
 }
 
 const LocaleMenuItem = (props: DropdownMenuItemProps) => {
-  const { t } = useTranslation()
-  const { icon, isoCode, onSelect } = props
+  const { fullName, icon, isoCode, onSelect } = props
 
   return (
     <DropdownMenu.Item asChild>
@@ -30,7 +27,7 @@ const LocaleMenuItem = (props: DropdownMenuItemProps) => {
       >
         <img alt={`${isoCode} flag`} className={s.dropItemFlag} src={icon} />
         <Typography className={s.dropdownText} variant={'caption'}>
-          {t(l.localeSwitcher.ownLanguages[isoCode as LangType])}
+          {fullName}
         </Typography>
       </div>
     </DropdownMenu.Item>
