+++ src/features/decks/ui/DecksPage.module.scss
@@ -1,0 +1,187 @@
+@import 'src/styles/mixins';
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+
+  to {
+    opacity: 1;
+  }
+}
+
+.common {
+  animation: fadeIn 0.6s;
+}
+
+.heading {
+  display: grid;
+  grid-template-rows: auto 1fr;
+  gap: 24px;
+  width: 100%;
+}
+
+.headingFirstRow {
+  display: flex;
+  justify-content: space-between;
+}
+
+.headingName {
+  display: flex;
+  flex-flow: row nowrap;
+  gap: 5px;
+  align-items: center;
+}
+
+.filters {
+  position: relative;
+
+  display: flex;
+  flex-flow: row nowrap;
+  gap: 25px;
+  align-items: end;
+
+  margin-bottom: 36px;
+
+  @media (width <= 860px) {
+    flex-flow: column nowrap;
+    align-items: center;
+  }
+}
+
+.tabsContainer {
+  position: relative;
+}
+
+.stepTabFilter {
+  width: 100%;
+}
+
+.countsFav {
+  position: absolute;
+  top: 20px;
+  right: -10px;
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  width: 20px;
+  height: 20px;
+
+  background-color: var(--color-accent-500);
+  border-radius: 50%;
+
+  p {
+    color: white;
+  }
+}
+
+.sliderBox {
+  display: flex;
+  flex-flow: column nowrap;
+  gap: 7px;
+
+  width: 100%;
+  max-width: 250px;
+
+  //.slider {
+  //  max-width: 150px;
+  //
+  //  @media (width <= 767px) {
+  //    max-width: 100%;
+  //  }
+  //}
+  //
+  //@media (width <= 767px) {
+  //  max-width: 100%;
+  //}
+}
+
+.clearFilterBox {
+  width: max-content;
+
+  @media (width <= 860px) {
+    width: 100%;
+  }
+}
+
+.clearFilter {
+  display: flex;
+  gap: 5px;
+  align-items: center;
+  justify-content: space-between;
+
+  height: 36px;
+
+  white-space: nowrap;
+
+  @media (width <= 860px) {
+    justify-content: center;
+    width: 100%;
+    max-width: 100%;
+  }
+}
+
+.arrow {
+  width: 24px;
+  height: 24px;
+  transition: transform 0.5s;
+}
+
+.rotate {
+  transform: rotate(180deg);
+}
+
+.input {
+  width: 100%;
+  max-width: 500px;
+
+  @media (width <= 1279px) {
+    max-width: 100%;
+  }
+}
+
+.footer {
+  width: 100%;
+  margin: 24px 0;
+}
+
+.empty {
+  text-align: center;
+}
+
+.boxForInput {
+  width: 100%;
+}
+
+.boxForSlider {
+  display: flex;
+  flex-flow: column nowrap;
+  gap: 7px;
+  width: 100%;
+}
+
+.buttonHelp {
+  overflow: hidden;
+
+  width: 24px;
+  height: 24px;
+  margin-left: 3px;
+  padding: 0;
+
+  background-color: transparent;
+  border: 1.5px solid var(--color-info-100);
+  border-radius: 50%;
+
+  &:hover {
+    background-color: transparent;
+    box-shadow: 0 0 8px 6px rgb(128 255 191 / 80%);
+  }
+
+  img {
+    width: 100%;
+    height: 100%;
+    border-radius: 50%;
+  }
+}
