+++ src/router/router.tsx
@@ -8,11 +8,12 @@ import { ForgotPassword } from '@/Pages/ForgotPassword/ForgotPassword'
 import { SignInPage } from '@/Pages/SignInPage/SignIn.page'
 import SignUp from '@/Pages/SignUp/SignUp'
 //import { CardsPage } from '@/Pages/CardsPage/Cards.page'
-import { DecksPage } from '@/Pages/DecksPage/Decks.page'
+
 import ErrorPage from '@/Pages/ErrorPage/ErrorPage'
 import { Layout } from '@/components/Layout/Layout'
 import Loading from '@/components/ui/Loading/Loading'
 import CardsPage from '@/features/cards/ui/Cards/cards.async'
+import { DecksPage } from '@/features/decks/ui/Decks.page'
 // import { Cards } from '@/features/cards/ui/Cards'
 import { LearnPage } from '@/features/learn/ui/LearnPage'
 import { ProfilePage } from '@/features/profile/ui/ProfilePage'
